% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_distance.R
\name{colour_blind_distance}
\alias{colour_blind_distance}
\title{Calculate a distance matrix between colours for colour blind people}
\usage{
colour_blind_distance(
  colours,
  deutan = TRUE,
  protan = FALSE,
  tritan = FALSE,
  gray = FALSE,
  fun = min,
  method = "emd"
)
}
\arguments{
\item{colours}{a vector of colours}

\item{deutan}{correct for deuteranomaly (red - green colour blindness).
Defaults to \code{TRUE}.}

\item{protan}{correct for protanopia (red - green colour blindness).
Defaults to \code{FALSE}.}

\item{tritan}{correct for tritanopia (blue - yellow colour blindness).
Defaults to \code{FALSE}.}

\item{gray}{correct for monochromacy (unable to distinguish colours).
Defaults to \code{FALSE}.}

\item{fun}{Function to aggregate different colour distance matrices.}

\item{method}{passed to \code{colordistance::getColorDistanceMatrix()}}
}
\description{
The function calculates the distance matrices for every option.
Then aggregates the matrices into a single distance matrix.
}
