% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_distance.R
\name{colour_distance}
\alias{colour_distance}
\title{Calculate a distance matrix between colours}
\usage{
colour_distance(colours, method = "emd")
}
\arguments{
\item{colours}{a vector of colours}

\item{method}{passed to \code{colordistance::getColorDistanceMatrix()}}
}
\description{
Calculate a distance matrix between colours
}
