% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_palette.R
\name{demo_palette}
\alias{demo_palette}
\title{Demonstrate colours in a palette}
\usage{
demo_palette(colours)
}
\arguments{
\item{colours}{a vector of colours}
}
\description{
Demonstrate colours in a palette
}
