% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{inbo.2015.colours}
\alias{inbo.2015.colours}
\alias{switchColour}
\alias{vlaanderen.2015.colours}
\title{Deprecated functions}
\usage{
inbo.2015.colours(n)

switchColour(new_colour = inbo_steun_blauw)

vlaanderen.2015.colours(n)
}
\arguments{
\item{n}{The number of colours}

\item{new_colour}{The new default colour.}
}
\description{
These functions will be removed from \code{INBOtheme} in the future.
}
