% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_en_data.R
\docType{data}
\name{dimensions}
\alias{dimensions}
\alias{page.height}
\alias{page.width}
\alias{column.width}
\alias{medium.width}
\title{Standard dimensions for the Elsevier style guide}
\description{
Several standard dimensions for the Elsevier style guide
-\code{page.height} Maximal height of a figuur (in inch)
-\code{page.width} Width of a figure covering two columns (in inch)
-\code{column.width} Width of a figure covering one column (in inch)
-\code{medium.width} Width of a figure covering 1.5 columns (in inch)
}
\seealso{
\code{\link{theme_elsevier}}
}
\concept{utils}
\keyword{datasets}
