% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inbo.2014.colours}
\alias{inbo.2014.colours}
\title{A colour pallet according the INBO style guide version <= 2014.}
\usage{
inbo.2014.colours(n)
}
\arguments{
\item{n}{The number of colours}
}
\value{
a vector of n hexadecimal values defining the colours.
}
\description{
A colour pallet according the INBO style guide version <= 2014.
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}

