% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{inbo.hoofd}
\alias{inbo.achtergrond}
\alias{inbo.bruinrood}
\alias{inbo.donkergroen}
\alias{inbo.geelgr}
\alias{inbo.grijs}
\alias{inbo.grijsblauw}
\alias{inbo.groen}
\alias{inbo.hoofd}
\alias{inbo.lichtblauw}
\alias{inbo.lichtgrijs}
\alias{inbo.lichtgroen}
\alias{inbo.munt}
\alias{inbo.oranje}
\alias{inbo.rood}
\alias{inbo.steun.blauw}
\alias{inbo.steun.donkerroos}
\alias{inbo.steun.geelgroen}
\title{Colour according to the INBO style guide version > 2014}
\usage{
inbo.hoofdkleur
}
\description{
Colour according to the INBO style guide as hexadecimal values.
\itemize{
 \item inbo.hoofd
 \item inbo.steun.donkerroos
 \item inbo.steun.blauw
 \item inbo.steun.geelgroen
 \item inbo.oranje
 \item inbo.rood
 \item inbo.bruinrood
 \item inbo.grijs
 \item inbo.lichtgrijs
 \item inbo.grijsblauw
 \item inbo.lichtblauw
 \item inbo.achtergrond
 \item inbo.munt
 \item inbo.lichtgroen
 \item inbo.donkergroen
 \item inbo.groen
 \item inbo.geelgr
}
}
\seealso{
\code{\link{theme_inbo2014}}
}
\keyword{datasets}

