% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_en_data.R
\docType{data}
\name{inbo_colours_2014}
\alias{inbo_colours_2014}
\alias{INBOblue}
\alias{INBObrown}
\alias{INBOdarkblue}
\alias{INBOdarkgreen}
\alias{INBOgreen}
\alias{INBOred}
\alias{INBOreddishbrown}
\alias{INBOextra}
\title{Colour according to the INBO style guide version <= 2014}
\description{
Colours according to the INBO style guide as hexadecimal values.
\itemize{
\item INBOblue
\item INBOdarkblue
\item INBOgreen
\item INBOdarkgreen
\item INBObrown
\item INBOreddishbrown
\item INBOred
\item INBOextra
}
}
\seealso{
\code{\link{theme_INBO}}

Other colours: 
\code{\link{INBO.colours}()},
\code{\link{demo_palette}()},
\code{\link{inbo.2015.colours}()},
\code{\link{inbo_colours_2015}},
\code{\link{percentcolor}()},
\code{\link{vlaanderen.2015.colours}()},
\code{\link{vlaanderen_colours}}
}
\author{
Thierry Onkelinx, Nicole De Groof
}
\concept{colours}
\keyword{datasets}
