% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_2015_colours.R
\name{inbo_palette}
\alias{inbo_palette}
\title{A Colour Palette According the INBO Style Guide Version >= 2015.}
\usage{
inbo_palette(n)
}
\arguments{
\item{n}{The number of colours}
}
\value{
a vector of n hexadecimal values defining the colours.
}
\description{
A Colour Palette According the INBO Style Guide Version >= 2015.
}
\seealso{
Other colours: 
\code{\link{demo_palette}()},
\code{\link{inbo_colours_2015}},
\code{\link{nara_palette}()},
\code{\link{percentcolor}()},
\code{\link{traffic_palette}()},
\code{\link{vlaanderen_colours}},
\code{\link{vlaanderen_palette}()}
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}
\concept{colours}
