% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{INBOblue}
\alias{INBOblue}
\alias{INBObrown}
\alias{INBOdarkblue}
\alias{INBOdarkgreen}
\alias{INBOextra}
\alias{INBOgreen}
\alias{INBOred}
\alias{INBOreddishbrown}
\title{Colour according to the INBO style guide version <= 2014}
\usage{
INBOblue
}
\description{
Colour according to the INBO style guide as hexadecimal values.
\itemize{
 \item INBOblue
 \item INBOdarkblue
 \item INBOgreen
 \item INBOdarkgreen
 \item INBObrown
 \item INBOreddishbrown
 \item INBOred
 \item INBOextra
}
}
\author{
Thierry Onkelinx, Nicole De Groof
}
\seealso{
\code{\link{theme_INBO}}
}
\keyword{datasets}

