% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{INBOtheme-package}
\alias{INBOtheme}
\alias{INBOtheme-package}
\title{Package with several ggplot2 themese}
\description{
Lod this package AFTER loading ggplot2. theme_INBO() will be set as the default theme.
}
\author{
Thierry Onkelinx \email{Thierry.Onkelinx@inbo.be}
}
\seealso{
\code{\link{theme_INBO}}, \code{\link{theme_elsevier}}
}
\keyword{package}

