% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_2015_colours.R
\name{nara_palette}
\alias{nara_palette}
\title{A Colour Palette for NARA reports.}
\usage{
nara_palette(n)
}
\arguments{
\item{n}{The number of colours}
}
\value{
a vector of n hexadecimal values defining the colours.
}
\description{
A Colour Palette for NARA reports.
}
\seealso{
Other colours: 
\code{\link{demo_palette}()},
\code{\link{inbo_hoofd}},
\code{\link{inbo_palette}()},
\code{\link{traffic_palette}()},
\code{\link{vl_yellow}},
\code{\link{vlaanderen_palette}()}
}
\concept{colours}
