% Generated by roxygen2 (4.0.2): do not edit by hand
\name{onLoad}
\alias{.onLoad}
\alias{onLoad}
\title{Set theme_INBO as default on loading the package}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{the name of the library}

\item{pkgname}{the name of the package}
}
\description{
Set theme_INBO as default on loading the package
}

