% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_distance.R
\name{optimal_order}
\alias{optimal_order}
\title{Order variables to maximise contrasts}
\usage{
optimal_order(distance, selected = character(0))
}
\arguments{
\item{distance}{a square distance matrix}

\item{selected}{optional vector of preselected variables}
}
\value{
a vector variable names
}
\description{
Start with variable which has the largest distance between all other
variables.
Then add the variable with the largest distance to the already selected
variables.
}
