% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_2015_colours.R
\name{ordinal_palette}
\alias{ordinal_palette}
\title{Palette for ordinal variables}
\usage{
ordinal_palette(n)
}
\arguments{
\item{n}{The number of colours}
}
\description{
The colour ramps depends on the active \code{ggplot2} theme.
}
