% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_palette.R
\name{show_palette}
\alias{show_palette}
\title{Show a palette on a single row}
\usage{
show_palette(colours)
}
\arguments{
\item{colours}{a vector of colours}
}
\description{
Show a palette on a single row
}
\seealso{
Other colours: 
\code{\link{demo_palette}()},
\code{\link{inbo_hoofd}},
\code{\link{inbo_palette}()},
\code{\link{nara_palette}()},
\code{\link{traffic_palette}()},
\code{\link{vl_yellow}},
\code{\link{vlaanderen_palette}()}
}
\concept{colours}
