% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_colour.R
\name{switch_colour}
\alias{switch_colour}
\title{Switch the default colour of all \code{ggplot2} \code{geoms}.}
\usage{
switch_colour(new_colour = inbo_steun_blauw)
}
\arguments{
\item{new_colour}{The new default colour.}
}
\description{
Switch the default colour of all \code{ggplot2} \code{geoms}.
}
\examples{
  switch_colour("black")
}
\seealso{
Other theme: 
\code{\link{theme_elsevier}()},
\code{\link{theme_inbo}()},
\code{\link{theme_map}()},
\code{\link{theme_nara}()},
\code{\link{theme_vlaanderen2015}()}
}
\concept{theme}
