% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_elsevier.R
\name{theme_elsevier}
\alias{theme_elsevier}
\title{Theme in compliance with the style guide of Elsevier journals}
\usage{
theme_elsevier(base_size = 7, base_family = "")
}
\arguments{
\item{base_size}{base fontsize}

\item{base_family}{base fonttype}
}
\description{
This theme is based on http://cdn.elsevier.com/assets/pdf_file/0010/109963/Artwork.pdf
}
\examples{
  library(ggplot2)
  p <- ggplot(mtcars, aes(x = mpg, y = drat)) + geom_point()
  p.elsevier <- p + theme_elsevier()
}
