% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_inbo2015.R
\name{theme_inbo}
\alias{theme_inbo}
\alias{theme_inbo2015}
\title{The theme in compliance with the INBO style guide version >= 2015.}
\usage{
theme_inbo(base_size = 12, base_family, transparent = FALSE)
}
\arguments{
\item{base_size}{Base fontsize}

\item{base_family}{Currently ignored and deprecated.}

\item{transparent}{Make backgrounds transparent.
\code{FALSE}: all backgrounds are white, \code{TRUE}: all backgrounds are transparent.
You can pass a vector to transparent.
In that case, it will check whether the values \code{"plot"}, \code{"panel"} and/or
\code{"legend"} are present.
The according items will be transparent.
Transparent panel will use grey instead of white gridlines.}
}
\description{
The theme in compliance with the INBO style guide version >= 2015.
}
\examples{
  library(ggplot2)
  p <- ggplot(mtcars, aes(x = mpg, y = drat)) + geom_point()
  p.inbo <- p + theme_inbo()
}
\seealso{
Other theme: 
\code{\link{switch_colour}()},
\code{\link{theme_elsevier}()},
\code{\link{theme_map}()},
\code{\link{theme_nara}()},
\code{\link{theme_vlaanderen2015}()}
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}
\concept{theme}
