% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theme_inbo2015}
\alias{theme_inbo2015}
\title{The theme in compliance with the INBO style guide version >= 2015.}
\usage{
theme_inbo2015(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base fontsize}

\item{base_family}{Base fonttype}
}
\description{
The theme in compliance with the INBO style guide version >= 2015.
}
\examples{
library(ggplot2)
  p <- ggplot(mtcars, aes(x = mpg, y = drat)) + geom_point()
  p.inbo <- p + theme_inbo2015()
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}

