% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map.R
\name{theme_map}
\alias{theme_map}
\title{a ggplot2 theme which removes labels, ticks and titles from both axes.}
\usage{
theme_map()
}
\description{
a ggplot2 theme which removes labels, ticks and titles from both axes.
}
\examples{
  library(ggplot2)
  p <- ggplot(mtcars, aes(x = mpg, y = drat)) + geom_point()
  p.map <- p + theme_map()
}
\seealso{
Other theme: 
\code{\link{switchColour}()},
\code{\link{theme_elsevier}()},
\code{\link{theme_inbo}()},
\code{\link{theme_vlaanderen2015}()}
}
\concept{theme}
