% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nara.R
\name{theme_nara}
\alias{theme_nara}
\title{The theme for NARA reports.}
\usage{
theme_nara(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Deprecated.
Use \code{theme_inbo()} instead.
}
\seealso{
Other theme: 
\code{\link{switch_colour}()},
\code{\link{theme_elsevier}()},
\code{\link{theme_inbo}()},
\code{\link{theme_map}()},
\code{\link{theme_vlaanderen2015}()}
}
\concept{theme}
