% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_vlaanderen2015.R
\name{theme_vlaanderen2015}
\alias{theme_vlaanderen2015}
\title{The theme in compliance with the Vlaanderen style guide version >= 2015.}
\usage{
theme_vlaanderen2015(base_size = 12, base_family = "", transparent = FALSE)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{transparent}{Make backgrounds transparent.
\code{FALSE}: all backgrounds are white, \code{TRUE}: all backgrounds are transparent.
You can pass a vector to transparent.
In that case, it will check whether the values \code{"plot"}, \code{"panel"} and/or
\code{"legend"} are present.
The according items will be transparent.
Transparent panel will use grey instead of white gridlines.}
}
\description{
The theme in compliance with the Vlaanderen style guide version >= 2015.
}
\examples{
  library(ggplot2)
  p <- ggplot(mtcars, aes(x = mpg, y = drat)) + geom_point()
  p.vl <- p + theme_vlaanderen2015()
}
\seealso{
Other theme: 
\code{\link{switch_colour}()},
\code{\link{theme_elsevier}()},
\code{\link{theme_inbo}()},
\code{\link{theme_map}()},
\code{\link{theme_nara}()}
}
\author{
Thierry Onkelinx, Nicole De Groof
}
\concept{theme}
