% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_en_data.R
\docType{data}
\name{vl.yellow}
\alias{vl.black}
\alias{vl.darkblue}
\alias{vl.darkbrown}
\alias{vl.darkgreen}
\alias{vl.darkred}
\alias{vl.darkyellow}
\alias{vl.grey1}
\alias{vl.grey2}
\alias{vl.grey3}
\alias{vl.grey4}
\alias{vl.grey5}
\alias{vl.grey6}
\alias{vl.grey7}
\alias{vl.lightblue}
\alias{vl.lightbrown}
\alias{vl.lightgreen}
\alias{vl.lightred}
\alias{vl.yellow}
\title{Some colours for the corporate style of vlaanderen.be. Taken from level 1, and level 2 pallette 7}
\usage{
vl.yellow
}
\description{
Colour according to the Vlaanderen style guide as hexadecimal values.
\itemize{
 \item vl.yellow
 \item vl.darkyellow
 \item vl.black
 \item vl.grey1
 \item vl.grey2
 \item vl.grey3
 \item vl.grey4
 \item vl.grey5
 \item vl.grey6
 \item vl.grey7
 \item vl.lightgreen
 \item vl.darkgreen
 \item vl.lightblue
 \item vl.darkblue
 \item vl.lightred
 \item vl.darkred
 \item vl.lightbrown
 \item vl.darkbrown
}
}
\author{
Thierry Onkelinx
}
\references{
http://webstijlgids.vlaanderen.be/element/kleurgebruik
}
\keyword{datasets}

