% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlaanderen_2015_colours.R
\name{vlaanderen.2015.colours}
\alias{vlaanderen.2015.colours}
\title{A colour pallet according the Vlaanderen style guide version >= 2015.}
\usage{
vlaanderen.2015.colours(n)
}
\arguments{
\item{n}{The number of colours}
}
\value{
a vector of n hexadecimal values defining the colours.
}
\description{
A colour pallet according the Vlaanderen style guide version >= 2015.
}
\references{
http://webstijlgids.vlaanderen.be/element/kleurgebruik
}
\seealso{
Other colours: 
\code{\link{INBO.colours}()},
\code{\link{demo_palette}()},
\code{\link{inbo.2015.colours}()},
\code{\link{inbo_colours_2014}},
\code{\link{inbo_colours_2015}},
\code{\link{percentcolor}()},
\code{\link{vlaanderen_colours}}
}
\author{
Thierry Onkelinx
}
\concept{colours}
