% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_and_data.R
\docType{data}
\name{vlaanderen_colours}
\alias{vlaanderen_colours}
\alias{vl_yellow}
\alias{vl_darkyellow}
\alias{vl_black}
\alias{vl_grey1}
\alias{vl_grey2}
\alias{vl_grey3}
\alias{vl_grey4}
\alias{vl_grey5}
\alias{vl_grey6}
\alias{vl_grey7}
\alias{vl_lightgreen}
\alias{vl_darkgreen}
\alias{vl_lightblue}
\alias{vl_darkblue}
\alias{vl_lightred}
\alias{vl_darkred}
\alias{vl_lightbrown}
\alias{vl_darkbrown}
\title{Some colours for the corporate style of vlaanderen.be.
Taken from level 1, and level 2 pallette 7}
\description{
Colour according to the Vlaanderen style guide as hexadecimal
values.
\itemize{
\item vl_yellow
\item vl_darkyellow
\item vl_black
\item vl_grey1
\item vl_grey2
\item vl_grey3
\item vl_grey4
\item vl_grey5
\item vl_grey6
\item vl_grey7
\item vl_lightgreen
\item vl_darkgreen
\item vl_lightblue
\item vl_darkblue
\item vl_lightred
\item vl_darkred
\item vl_lightbrown
\item vl_darkbrown
}
}
\references{
http://webstijlgids.vlaanderen.be/element/kleurgebruik
}
\seealso{
Other colours: 
\code{\link{demo_palette}()},
\code{\link{inbo_colours_2015}},
\code{\link{inbo_palette}()},
\code{\link{nara_palette}()},
\code{\link{percentcolor}()},
\code{\link{vlaanderen_palette}()}
}
\author{
Thierry Onkelinx
}
\concept{colours}
\keyword{datasets}
