% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlaanderen_2015_colours.R
\name{vlaanderen_palette}
\alias{vlaanderen_palette}
\title{A colour pallet according the Vlaanderen style guide version >= 2015.}
\usage{
vlaanderen_palette(n)
}
\arguments{
\item{n}{The number of colours}
}
\value{
a vector of n hexadecimal values defining the colours.
}
\description{
A colour pallet according the Vlaanderen style guide version >= 2015.
}
\references{
http://webstijlgids.vlaanderen.be/element/kleurgebruik
}
\seealso{
Other colours: 
\code{\link{demo_palette}()},
\code{\link{inbo_colours_2015}},
\code{\link{inbo_palette}()},
\code{\link{nara_palette}()},
\code{\link{percentcolor}()},
\code{\link{traffic_palette}()},
\code{\link{vlaanderen_colours}}
}
\author{
Thierry Onkelinx
}
\concept{colours}
