% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_palette.R
\name{demo_palette}
\alias{demo_palette}
\title{Demonstrate colours in a palette}
\usage{
demo_palette(colours)
}
\arguments{
\item{colours}{a vector of colours}
}
\description{
Demonstrate colours in a palette
}
\seealso{
Other colours: 
\code{\link{INBO.colours}()},
\code{\link{inbo.2015.colours}()},
\code{\link{inbo_colours_2014}},
\code{\link{inbo_colours_2015}},
\code{\link{percentcolor}()},
\code{\link{vlaanderen.2015.colours}()},
\code{\link{vlaanderen_colours}}
}
\concept{colours}
