% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_palette.R
\name{demo_palette}
\alias{demo_palette}
\title{Demonstrate colours in a palette}
\usage{
demo_palette(colours)
}
\arguments{
\item{colours}{a vector of colours}
}
\description{
Create an image with a number of rectangles coloured along the input vector.
}
\seealso{
Other colours: 
\code{\link{inbo_hoofd}},
\code{\link{inbo_palette}()},
\code{\link{nara_palette}()},
\code{\link{traffic_palette}()},
\code{\link{vl_yellow}},
\code{\link{vlaanderen_palette}()}
}
\concept{colours}
