% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_and_data.R
\docType{data}
\name{dimensions}
\alias{dimensions}
\alias{page_height}
\alias{page_width}
\alias{column_width}
\alias{medium_width}
\title{Standard dimensions for the Elsevier style guide}
\description{
Several standard dimensions for the Elsevier style guide
-\code{page_height} Maximal height of a figuur (in inch)
-\code{page_width} Width of a figure covering two columns (in inch)
-\code{column_width} Width of a figure covering one column (in inch)
-\code{medium_width} Width of a figure covering 1.5 columns (in inch)
}
\seealso{
\code{\link{theme_elsevier}}
}
\concept{utils}
\keyword{datasets}
