% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inbo.2015.colours}
\alias{inbo.2015.colours}
\title{A colour pallet according the INBO style guide version >= 2015.}
\usage{
inbo.2015.colours(n)
}
\arguments{
\item{n}{The number of colours}
}
\value{
a vector of n hexadecimal values defining the colours.
}
\description{
A colour pallet according the INBO style guide version >= 2015.
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}

