% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_2015_colours.R
\name{inbo.2015.colours}
\alias{inbo.2015.colours}
\title{A colour pallet according the INBO style guide version >= 2015.}
\usage{
inbo.2015.colours(n)
}
\arguments{
\item{n}{The number of colours}
}
\value{
a vector of n hexadecimal values defining the colours.
}
\description{
A colour pallet according the INBO style guide version >= 2015.
}
\seealso{
Other colours: \code{\link{INBO.colours}},
  \code{\link{demo_palette}},
  \code{\link{inbo_colours_2014}},
  \code{\link{inbo_colours_2015}},
  \code{\link{percentcolor}},
  \code{\link{vlaanderen.2015.colours}},
  \code{\link{vlaanderen_colours}}
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}
\concept{colours}
