% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INBOcolours.R
\name{INBO.colours}
\alias{INBO.colours}
\title{A colour pallet according the INBO style guide version <= 2014.}
\usage{
INBO.colours(n)
}
\arguments{
\item{n}{The number of colours}
}
\value{
a vector of n hexadecimal values defining the colours.
}
\description{
A colour pallet according the INBO style guide version <= 2014.
}
\seealso{
Other colours: \code{\link{demo_palette}},
  \code{\link{inbo.2015.colours}},
  \code{\link{inbo_colours_2014}},
  \code{\link{inbo_colours_2015}},
  \code{\link{percentcolor}},
  \code{\link{vlaanderen.2015.colours}},
  \code{\link{vlaanderen_colours}}
}
\concept{colours}
