\name{INBO.colours}
\alias{INBO.colours}
\title{A colour pallet according the INBO style guide.}
\usage{
INBO.colours(n)
}
\arguments{
  \item{n}{The number of colours}
}
\value{
a vector of n hexadecimal values defining the colours.
}
\description{
A colour pallet according the INBO style guide.
}

