% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_and_data.R
\docType{data}
\name{inbo_colours_2015}
\alias{inbo_colours_2015}
\alias{inbo_hoofd}
\alias{inbo_steun_donkerroos}
\alias{inbo_steun_blauw}
\alias{inbo_steun_geelgroen}
\alias{inbo_oranje}
\alias{inbo_rood}
\alias{inbo_bruinrood}
\alias{inbo_grijs}
\alias{inbo_lichtgrijs}
\alias{inbo_grijsblauw}
\alias{inbo_lichtblauw}
\alias{inbo_achtergrond}
\alias{inbo_munt}
\alias{inbo_lichtgroen}
\alias{inbo_donkergroen}
\alias{inbo_groen}
\alias{inbo_geelgr}
\title{Colour according to the INBO style guide version >= 2015}
\description{
Colour according to the INBO style guide as hexadecimal values.
\itemize{
\item inbo_hoofd
\item inbo_steun_donkerroos
\item inbo_steun_blauw
\item inbo_steun_geelgroen
\item inbo_oranje
\item inbo_rood
\item inbo_bruinrood
\item inbo_grijs
\item inbo_lichtgrijs
\item inbo_grijsblauw
\item inbo_lichtblauw
\item inbo_achtergrond
\item inbo_munt
\item inbo_lichtgroen
\item inbo_donkergroen
\item inbo_groen
\item inbo_geelgr
}
}
\seealso{
\code{\link{theme_inbo2015}}

Other colours: 
\code{\link{demo_palette}()},
\code{\link{inbo_palette}()},
\code{\link{percentcolor}()},
\code{\link{vlaanderen_colours}},
\code{\link{vlaanderen_palette}()}
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}
\concept{colours}
\keyword{datasets}
