% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_and_data.R
\docType{data}
\name{inbo_hoofd}
\alias{inbo_hoofd}
\alias{inbo_steun_donkerroos}
\alias{inbo_steun_blauw}
\alias{inbo_steun_geelgroen}
\alias{inbo_oranje}
\alias{inbo_rood}
\alias{inbo_bruinrood}
\alias{inbo_grijs}
\alias{inbo_lichtgrijs}
\alias{inbo_grijsblauw}
\alias{inbo_lichtblauw}
\alias{inbo_achtergrond}
\alias{inbo_munt}
\alias{inbo_lichtgroen}
\alias{inbo_donkergroen}
\alias{inbo_groen}
\alias{inbo_geelgr}
\title{Colour according to the INBO style guide version >= 2015}
\description{
Colour according to the INBO style guide as hexadecimal values.
\itemize{
\item \code{inbo_hoofd}
\item \code{inbo_steun_donkerroos}
\item \code{inbo_steun_blauw}
\item \code{inbo_steun_geelgroen}
\item \code{inbo_oranje}
\item \code{inbo_rood}
\item \code{inbo_bruinrood}
\item \code{inbo_grijs}
\item \code{inbo_lichtgrijs}
\item \code{inbo_grijsblauw}
\item \code{inbo_lichtblauw}
\item \code{inbo_achtergrond}
\item \code{inbo_munt}
\item \code{inbo_lichtgroen}
\item \code{inbo_donkergroen}
\item \code{inbo_groen}
\item \code{inbo_geelgr}
}
}
\seealso{
\code{theme_inbo2015()}

Other colours: 
\code{\link{demo_palette}()},
\code{\link{inbo_palette}()},
\code{\link{nara_palette}()},
\code{\link{traffic_palette}()},
\code{\link{vl_yellow}},
\code{\link{vlaanderen_palette}()}
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}
\concept{colours}
\keyword{datasets}
