% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_en_data.R
\docType{package}
\name{INBOtheme-package}
\alias{INBOtheme-package}
\alias{INBOtheme}
\title{Package with several ggplot2 themese}
\description{
Load this package AFTER loading ggplot2. theme_inbo() will be set as the default theme.
}
\seealso{
\code{\link{theme_inbo}}, \code{\link{theme_elsevier}}
}
\author{
Thierry Onkelinx \email{Thierry.Onkelinx@inbo.be}
}
\keyword{package}
