% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{onLoad}
\alias{.onLoad}
\alias{onLoad}
\title{Set theme_inbo2015 as default on loading the package}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{the name of the library}

\item{pkgname}{the name of the package}
}
\description{
Set theme_inbo2015 as default on loading the package
}

