% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package_en_data.R
\docType{data}
\name{page.height}
\alias{column.width}
\alias{medium.width}
\alias{page.height}
\alias{page.width}
\title{Standard dimensions for the Elsevier style guide
\itemize{
 \item{page.height}{Maximal height of a figuur (in inch)}
 \item{page.width}{Width of a figure covering two columns (in inch)}
 \item{column.width}{Width of a figure covering one column (in inch)}
 \item{medium.width}{Width of a figure covering 1.5 columns (in inch)}
}}
\usage{
page.height
}
\description{
Standard dimensions for the Elsevier style guide
\itemize{
 \item{page.height}{Maximal height of a figuur (in inch)}
 \item{page.width}{Width of a figure covering two columns (in inch)}
 \item{column.width}{Width of a figure covering one column (in inch)}
 \item{medium.width}{Width of a figure covering 1.5 columns (in inch)}
}
}
\seealso{
\code{\link{theme_elsevier}}
}
\keyword{datasets}

