\docType{data}
\name{page.height}
\alias{column.width,}
\alias{medium.width}
\alias{page.height}
\alias{page.width,}
\title{Standard dimensions for theElsevier style guide
\itemize{
 \item{page.height}{Maximal height of a figuur (in inch)}
 \item{page.width}{With of a figure covering two columns (in inch)}
 \item{column.width}{With of a figure covering one column (in inch)}
 \item{medium.width}{With of a figure covering 1.5 columns (in inch)}
}}
\usage{
page.height
}
\description{
Standard dimensions for theElsevier style guide \itemize{
\item{page.height}{Maximal height of a figuur (in inch)}
\item{page.width}{With of a figure covering two columns (in
inch)} \item{column.width}{With of a figure covering one
column (in inch)} \item{medium.width}{With of a figure
covering 1.5 columns (in inch)} }
}
\seealso{
\code{\link{theme_elsevier}}
}
\keyword{datasets}

