% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentcolor.R
\name{percentcolor}
\alias{percentcolor}
\title{Make a colour lighter}
\usage{
percentcolor(hex, percent = 0.95)
}
\arguments{
\item{hex}{Hexadecimal values of the colour}

\item{percent}{The percentage to lighten the colour. 0 = no change, 1 = pure white}
}
\description{
Make a colour lighter
}

