% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_inbo2015.R
\name{margin}
\alias{margin}
\title{Re-exported Function From \code{ggplot2}}
\usage{
margin(t = 0, r = 0, b = 0, l = 0, unit = "pt")
}
\arguments{
\item{t, r, b, l}{Dimensions of each margin. (To remember order, think trouble).}

\item{unit}{Default units of dimensions. Defaults to "pt" so it
can be most easily scaled with the text.}
}
\description{
See \code{\link[ggplot2:element]{ggplot2::margin()}} in \code{ggplot2}.
}
