% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_discrete}
\alias{scale_colour_discrete}
\alias{scale_color_discrete}
\title{redefine \code{ggplot2::scale_colour_discrete()}}
\usage{
scale_colour_discrete(..., type)

scale_color_discrete(..., type)
}
\arguments{
\item{...}{Additional parameters passed on to the scale type,}

\item{type}{Ignored argument.
Only present to match the arguments of \code{ggplot2::scale_fill_discrete()}}
}
\description{
Selects the colours depending on the current set theme.
}
\seealso{
Other scales: 
\code{\link{scale_colour_gradient2}()},
\code{\link{scale_colour_gradient}()},
\code{\link{scale_colour_viridis_d}()},
\code{\link{scale_fill_discrete}()},
\code{\link{scale_fill_gradient2}()},
\code{\link{scale_fill_gradient}()}
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}
\concept{scales}
