% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_gradient}
\alias{scale_colour_gradient}
\title{redefine scale_colour_gradient}
\usage{
scale_colour_gradient(..., low = INBOblue, high = INBOred, space = "Lab",
  na.value = "grey50", guide = "colourbar")
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{continuous_scale}}
to control name, limits, breaks, labels and so forth.}

\item{low}{Colour for the low end of the gradient}

\item{high}{Colour for the high end of the gradient}

\item{space}{colour space in which to calculate gradient. Must be "Lab" -
other values are deprecated.}

\item{na.value}{Colour to use for missing values}

\item{guide}{Type of legend. Use \code{"colourbar"} for continuous
colour bar, or \code{"legend"} for discrete colour legend.}
}
\description{
Selects the colours from the INBO style guide when the default text colour is INBObrown.
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}

