% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_gradient2}
\alias{scale_colour_gradient2}
\title{redefine scale_colour_gradient2}
\usage{
scale_colour_gradient2(..., low = INBOblue, high = INBOred, mid = "white",
  midpoint = 0, space = "Lab", na.value = "grey50", guide = "colourbar")
}
\description{
Selects the colours from the INBO style guide when the default text colour is INBObrown.
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}

