% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_discrete}
\alias{scale_fill_discrete}
\title{redefine scale_fill_discrete}
\usage{
scale_fill_discrete(..., type)
}
\arguments{
\item{...}{Additional parameters passed on to the scale type,}

\item{type}{One of the following:
\itemize{
\item A character vector of color codes. The codes are used for a 'manual' color
scale as long as the number of codes exceeds the number of data levels
(if there are more levels than codes, \code{\link[ggplot2:scale_colour_hue]{scale_colour_hue()}}/\code{\link[ggplot2:scale_fill_hue]{scale_fill_hue()}}
are used to construct the default scale). If this is a named vector, then the color values
will be matched to levels based on the names of the vectors. Data values that
don't match will be set as \code{na.value}.
\item A list of character vectors of color codes. The minimum length vector that exceeds the
number of data levels is chosen for the color scaling. This is useful if you
want to change the color palette based on the number of levels.
\item A function that returns a discrete colour/fill scale (e.g., \code{\link[ggplot2:scale_fill_hue]{scale_fill_hue()}},
\code{\link[ggplot2:scale_fill_brewer]{scale_fill_brewer()}}, etc).
}}
}
\description{
Selects the colours depending on the current set theme.
}
\seealso{
Other scales: 
\code{\link{scale_colour_discrete}()},
\code{\link{scale_colour_gradient2}()},
\code{\link{scale_colour_gradient}()},
\code{\link{scale_colour_viridis_d}()},
\code{\link{scale_fill_gradient2}()},
\code{\link{scale_fill_gradient}()}
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}
\concept{scales}
