% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scale_fill_discrete}
\alias{scale_fill_discrete}
\title{redefine scale_cfill_discrete}
\usage{
scale_fill_discrete(..., h = c(0, 360) + 15, c = 100, l = 65,
  h.start = 0, direction = 1, na.value = "grey50")
}
\arguments{
\item{...}{other arguments passed to discrete_scale}

\item{h}{range of hues to use, in [0, 360]}

\item{c}{chroma (intensity of colour), maximum value varies depending on}

\item{l}{luminance (lightness), in [0, 100]}

\item{h.start}{hue to start at}

\item{direction}{direction to travel around the colour wheel, 1 = clockwise, -1 = counter-clockwise}

\item{na.value}{colour for missing values}
}
\description{
Selects the colour from the INBO style guide when the default text colour is INBObrown.
}

