% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_discrete}
\alias{scale_fill_discrete}
\title{redefine scale_fill_discrete}
\usage{
scale_fill_discrete(
  ...,
  h = c(0, 360) + 15,
  c = 100,
  l = 65,
  h.start = 0,
  direction = 1,
  na.value = "grey50",
  aesthetics = "fill"
)
}
\arguments{
\item{...}{Additional parameters passed on to the scale type,}
}
\description{
Selects the colours from the INBO style guide when the default text colour is \code{\link{inbo.steun.donkerroos}} or \code{\link{vl.black}}. The former is the default when \code{\link{theme_inbo}} is set as the default theme. The latter occurs when \code{\link{theme_vlaanderen2015}} is set as default theme.
}
\seealso{
Other scales: 
\code{\link{scale_colour_discrete}()},
\code{\link{scale_colour_gradient2}()},
\code{\link{scale_colour_gradient}()},
\code{\link{scale_fill_gradient2}()},
\code{\link{scale_fill_gradient}()}
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}
\concept{scales}
