% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scale_fill_gradient}
\alias{scale_fill_gradient}
\title{redefine scale_fill_gradient}
\usage{
scale_fill_gradient(..., low = INBOblue, high = INBOred, space = "Lab",
  na.value = "grey50", guide = "colourbar")
}
\arguments{
\item{...}{other arguments passed to continuous_scale}

\item{low}{colour for lowest value}

\item{high}{colour for highest value}

\item{space}{method to change the colour gradually}

\item{na.value}{colour for missing values}

\item{guide}{type of legend}
}
\description{
Selects the colours from the INBO style guide when the default text colour is INBObrown.
}
\author{
Thierry Onkelinx, Oona Op de Weerdt, Nicole De Groof
}

