% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_colour.R
\name{switchColour}
\alias{switchColour}
\title{Switch the default colour of all geoms}
\usage{
switchColour(new.colour = inbo.steun.blauw)
}
\arguments{
\item{new.colour}{the new default colour}
}
\description{
Switch the default colour of all geoms
}
\examples{
  switchColour("black")
}
\seealso{
Other theme: 
\code{\link{theme_elsevier}()},
\code{\link{theme_inbo}()},
\code{\link{theme_map}()},
\code{\link{theme_vlaanderen2015}()}
}
\concept{theme}
