% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/switch_colour.R
\name{switchColour}
\alias{switchColour}
\title{Switch the default colour of all geoms}
\usage{
switchColour(new.colour = inbo.steun.geelgroen)
}
\arguments{
\item{new.colour}{the new default colour}
}
\description{
Switch the default colour of all geoms
}
\examples{
switchColour("black")
}

