\name{theme_elsevier}
\alias{theme_elsevier}
\title{Theme in compliance with the style guide of Elsevier journals}
\usage{
theme_elsevier(base_size = 7, base_family = "")
}
\arguments{
  \item{base_size}{base fontsize}

  \item{base_family}{base fonttype}
}
\description{
This theme is base on
http://cdn.elsevier.com/assets/pdf_file/0010/109963/Artwork.pdf
}

