% Generated by roxygen2 (4.0.1): do not edit by hand
\name{theme_INBO}
\alias{theme_INBO}
\title{The theme in compliance with the INBO style guide, version <= 2014.}
\usage{
theme_INBO(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base fontsize}

\item{base_family}{Base fonttype}
}
\description{
The theme in compliance with the INBO style guide, version <= 2014.
}
\examples{
library(ggplot2)
  p <- ggplot(mtcars, aes(x = mpg, y = drat)) + geom_point()
  p.inbo <- p + theme_INBO()
}

