\name{theme_INBO}
\alias{theme_INBO}
\title{The theme in compliance with the INBO style guide.}
\usage{
theme_INBO(base_size = 12, base_family = "")
}
\arguments{
  \item{base_size}{Base fontsize}

  \item{base_family}{Base fonttype}
}
\description{
The theme in compliance with the INBO style guide.
}

