% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nara.R
\name{theme_nara}
\alias{theme_nara}
\title{The theme for NARA reports.}
\usage{
theme_nara(base_size = 12)
}
\arguments{
\item{base_size}{Base fontsize}
}
\description{
The theme for NARA reports.
}
\examples{
  library(ggplot2)
  p <- ggplot(mtcars, aes(x = mpg, y = drat)) + geom_point()
  p.inbo <- p + theme_nara()
}
\seealso{
Other theme: 
\code{\link{switch_colour}()},
\code{\link{theme_elsevier}()},
\code{\link{theme_inbo}()},
\code{\link{theme_map}()},
\code{\link{theme_vlaanderen2015}()}
}
\concept{theme}
