% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_and_data.R
\docType{data}
\name{vl_yellow}
\alias{vl_yellow}
\alias{vl_darkyellow}
\alias{vl_black}
\alias{vl_grey1}
\alias{vl_grey2}
\alias{vl_grey3}
\alias{vl_grey4}
\alias{vl_grey5}
\alias{vl_grey6}
\alias{vl_grey7}
\alias{vl_lightgreen}
\alias{vl_darkgreen}
\alias{vl_lightblue}
\alias{vl_darkblue}
\alias{vl_lightred}
\alias{vl_darkred}
\alias{vl_lightbrown}
\alias{vl_darkbrown}
\title{Some colours for the corporate style of vlaanderen.be.
Taken from level 1, and level 2 palette 7}
\description{
Colour according to the Flanders style guide as hexadecimal
values.
\itemize{
\item \code{vl_yellow}
\item \code{vl_darkyellow}
\item \code{vl_black}
\item \code{vl_grey1}
\item \code{vl_grey2}
\item \code{vl_grey3}
\item \code{vl_grey4}
\item \code{vl_grey5}
\item \code{vl_grey6}
\item \code{vl_grey7}
\item \code{vl_lightgreen}
\item \code{vl_darkgreen}
\item \code{vl_lightblue}
\item \code{vl_darkblue}
\item \code{vl_lightred}
\item \code{vl_darkred}
\item \code{vl_lightbrown}
\item \code{vl_darkbrown}
}
}
\references{
http://webstijlgids.vlaanderen.be/element/kleurgebruik
}
\seealso{
Other colours: 
\code{\link{demo_palette}()},
\code{\link{inbo_hoofd}},
\code{\link{inbo_palette}()},
\code{\link{nara_palette}()},
\code{\link{show_palette}()},
\code{\link{traffic_palette}()},
\code{\link{vlaanderen_palette}()}
}
\author{
Thierry Onkelinx
}
\concept{colours}
\keyword{datasets}
