% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_en_data.R
\docType{data}
\name{vlaanderen_colours}
\alias{vlaanderen_colours}
\alias{vl.yellow}
\alias{vl.darkyellow}
\alias{vl.black}
\alias{vl.grey1}
\alias{vl.grey2}
\alias{vl.grey3}
\alias{vl.grey4}
\alias{vl.grey5}
\alias{vl.grey6}
\alias{vl.grey7}
\alias{vl.lightgreen}
\alias{vl.darkgreen}
\alias{vl.lightblue}
\alias{vl.darkblue}
\alias{vl.lightred}
\alias{vl.darkred}
\alias{vl.lightbrown}
\alias{vl.darkbrown}
\title{Some colours for the corporate style of vlaanderen.be.
Taken from level 1, and level 2 pallette 7}
\description{
Colour according to the Vlaanderen style guide as hexadecimal
values.
\itemize{
\item vl.yellow
\item vl.darkyellow
\item vl.black
\item vl.grey1
\item vl.grey2
\item vl.grey3
\item vl.grey4
\item vl.grey5
\item vl.grey6
\item vl.grey7
\item vl.lightgreen
\item vl.darkgreen
\item vl.lightblue
\item vl.darkblue
\item vl.lightred
\item vl.darkred
\item vl.lightbrown
\item vl.darkbrown
}
}
\references{
http://webstijlgids.vlaanderen.be/element/kleurgebruik
}
\seealso{
Other colours: 
\code{\link{demo_palette}()},
\code{\link{inbo.2015.colours}()},
\code{\link{inbo_colours_2015}},
\code{\link{percentcolor}()},
\code{\link{vlaanderen.2015.colours}()}
}
\author{
Thierry Onkelinx
}
\concept{colours}
\keyword{datasets}
